refactor scaling
