--- shardingsphere-scaling/.gitignore
@@ -1,6 +1,0 @@
-# maven ignore
-target/
-*.iml
-
-# idea ignore
-.idea/
